backend/requirement.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
matplotlib
